make the tests work with the new RBF
